refactor(web): move to separate folder
